Feat: Fridge List -> Sort by Distance to Center Location
